Move config/types.ts out of user-facing /config folder
