+++ clients/privacy-center/types/config.ts
@@ -1,0 +1,31 @@
+export type Config = {
+  title: string;
+  description: string;
+  server_url_development?: string;
+  server_url_production?: string;
+  logo_path: string;
+  actions: PrivacyRequestOption[];
+  includeConsent: boolean;
+  consent?: {
+    cookieName?: string;
+    consentOptions: ConfigConsentOption[];
+  };
+};
+
+export type PrivacyRequestOption = {
+  policy_key: string;
+  icon_path: string;
+  title: string;
+  description: string;
+  identity_inputs: object;
+};
+
+export type ConfigConsentOption = {
+  cookieKeys: string[];
+  default?: boolean;
+  description: string;
+  fidesDataUseKey: string;
+  highlight?: boolean;
+  name: string;
+  url: string;
+};
